Adjust namespace in the config area.
