+++ osu.Game.Rulesets.Karaoke/Edit/Configs/Generator/Language/LanguageDetectorConfigDialog.cs
@@ -4,14 +4,21 @@
 #nullable disable
 
 using osu.Framework.Bindables;
+using osu.Game.Overlays;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Rulesets.Karaoke.Edit.Generator.Language;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Configs.Generator.Languages
+namespace osu.Game.Rulesets.Karaoke.Edit.Configs.Generator.Language
 {
-    public class LanguageDetectorConfigPopover : GeneratorConfigPopover<LanguageDetectorConfig>
+    public class LanguageDetectorConfigDialog : GeneratorConfigDialog<LanguageDetectorConfig>
     {
         protected override KaraokeRulesetEditGeneratorSetting Config => KaraokeRulesetEditGeneratorSetting.LanguageDetectorConfig;
+
+        protected override OverlayColourScheme OverlayColourScheme => OverlayColourScheme.Orange;
+
+        protected override string Title => "Language generator config";
+
+        protected override string Description => "Change config for language generator.";
 
         protected override GeneratorConfigSection[] CreateConfigSection(Bindable<LanguageDetectorConfig> current)
         {
