+++ osu.Game.Rulesets.Karaoke/Edit/Configs/Generator/Language/LanguageDetectorConfigPopover.cs
@@ -4,21 +4,14 @@
 #nullable disable
 
 using osu.Framework.Bindables;
-using osu.Game.Overlays;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Rulesets.Karaoke.Edit.Generator.Language;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Configs.Generator.Languages
+namespace osu.Game.Rulesets.Karaoke.Edit.Configs.Generator.Language
 {
-    public class LanguageDetectorConfigDialog : GeneratorConfigDialog<LanguageDetectorConfig>
+    public class LanguageDetectorConfigPopover : GeneratorConfigPopover<LanguageDetectorConfig>
     {
         protected override KaraokeRulesetEditGeneratorSetting Config => KaraokeRulesetEditGeneratorSetting.LanguageDetectorConfig;
-
-        protected override OverlayColourScheme OverlayColourScheme => OverlayColourScheme.Orange;
-
-        protected override string Title => "Language generator config";
-
-        protected override string Description => "Change config for language generator.";
 
         protected override GeneratorConfigSection[] CreateConfigSection(Bindable<LanguageDetectorConfig> current)
         {
